fix(core): align runtime transport and protocol parity
Normalize HTTP parity metadata and websocket-backed message/profile flows so runtime behavior matches the latest parity audit.

+++ AioTieba4DotNet/Models/Users/UserPost.cs
@@ -26,7 +26,7 @@ public class UserPost
     /// <summary>
     ///     回复id
     /// </summary>
-    public int Pid { get; set; }
+    public long Pid { get; set; }
 
     /// <summary>
     ///     是否为楼中楼
